TOEM
TOEM is the first Tokem backed by emerald cane in Brazil, which comes with the purpose of trading Gold and Precious Stones, such as Emerald, Diamond, among others. These commodities are recognized for their properties as a store of value and are the target of many investors for the manufacture of high luxury jewelry and decorative pieces. 

https://drive.google.com/file/d/1R5qwwrYAJzTamXhAv-3Qu6f0MQKJ3DRt/view

https://en.tokenemerald.com/

https://coinpaprika.com/coin/toem-toem-token-emerald/